Total reestructuration: now we use applications and (I hope) we can test models

+++ cleartest.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env DJANGO_SETTINGS_MODULE=settings python
 import django.contrib.auth.models as am
 
-try:
+#try:
+if 1:
     test_user = am.User.objects.get(username='test')
     print "User test found"
     test_user.delete()
     print "User test deleted"
-except am.DoesNotExist:
-    print "User test not found"
+#except:
+   # print "User test not found"
